move to helpers general_purpose_funcs Fixes #27
